machine learning/rl_starter/gemma/requirements.txt
torch>=2.1
transformers>=4.42
accelerate>=0.30
bitsandbytes>=0.43
peft>=0.11
datasets>=2.19
sentencepiece>=0.2
faiss-cpu>=1.8
tqdm>=4.66
matplotlib>=3.8
numpy>=1.24
scikit-learn>=1.4
